refactor: move WebClient mocks to index.helper.ts
