+++ functions/src/__tests__/index.test.ts
@@ -7,12 +7,13 @@ import {
   expect,
   jest,
 } from "@jest/globals";
-import {
-  MockedWebClient,
-  MockWebClient,
-} from "@slack-wrench/jest-mock-web-client";
 import * as admin from "firebase-admin";
-import { setupFunctionsTest, postSlackEvent, rotations } from "./index.helper";
+import {
+  setupFunctionsTest,
+  postSlackEvent,
+  rotations,
+  mockSlackWebClient,
+} from "./index.helper";
 
 const functionsTest = setupFunctionsTest();
 
@@ -25,49 +26,7 @@ describe("functions", () => {
     functionsTest.cleanup();
   });
 
-  // Mock Slack API client
-  let clients: MockWebClient[];
-  beforeEach(async () => {
-    clients = MockedWebClient.mock.instances;
-  });
-  afterEach(async () => {
-    clients.map((client) => {
-      client.chat.postMessage.mockClear();
-      client.chat.update.mockClear();
-      client.chat.postEphemeral.mockClear();
-    });
-  });
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  type IsAny<T> = (any extends T ? true : false) extends true ? true : false;
-
-  /**
-   * jest.Mock<any, any> である T の末端のプロパティを、ドット区切りの keyPath で列挙する
-   */
-  type KeyPath<T> = {
-    [K in keyof T]: K extends string
-      ? // @slack-wrench/jest-mock-web-client の参照する jest の型定義が古い関係で、
-        // 「jest.Mock<any, any> か否か」は「any か否か」で判別する必要がある
-        IsAny<T[K]> extends true
-        ? K // leaf
-        : T[K] extends Record<string, unknown>
-        ? `${K}.${KeyPath<T[K]>}`
-        : never
-      : never;
-  }[keyof T];
-
-  const getClientCalls = (keyPath: KeyPath<MockWebClient>) => {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const trace = (obj: any, keyPathParts: string[]): any => {
-      if (keyPathParts.length === 0) return obj;
-      const [head, ...rest] = keyPathParts;
-      return trace(obj[head], rest);
-    };
-
-    return clients.map(
-      (client) => trace(client, keyPath.split(".")).mock.calls
-    );
-  };
+  const { getSlackWebClientCalls } = mockSlackWebClient();
 
   // Firestore
   const rotationsRef = admin.firestore().collection("rotations");
@@ -96,7 +55,7 @@ describe("functions", () => {
           // ...more unused parameters
         });
         expect(res.body).toEqual({}); // ack
-        expect(getClientCalls("views.open")).toMatchSnapshot();
+        expect(getSlackWebClientCalls("views.open")).toMatchSnapshot();
 
         expect(await getAllRotations()).toEqual(rotations);
       });
@@ -224,7 +183,7 @@ describe("functions", () => {
           payload: JSON.stringify(payloadObj),
         });
         expect(res.body).toEqual({}); // ack
-        expect(getClientCalls("chat.postMessage")).toMatchSnapshot();
+        expect(getSlackWebClientCalls("chat.postMessage")).toMatchSnapshot();
 
         expect(await getAllRotations()).toEqual([
           submittedRotation,
@@ -244,7 +203,7 @@ describe("functions", () => {
           }),
         });
         expect(res.body).toEqual({}); // ack
-        expect(getClientCalls("chat.postMessage")).toMatchSnapshot();
+        expect(getSlackWebClientCalls("chat.postMessage")).toMatchSnapshot();
 
         expect(await getAllRotations()).toEqual([
           ...rotations.slice(0, 3),
@@ -305,7 +264,7 @@ describe("functions", () => {
             value: "rotate:rotation-1",
           });
           expect(res.body).toEqual({}); // ack
-          expect(getClientCalls("chat.update")).toMatchSnapshot();
+          expect(getSlackWebClientCalls("chat.update")).toMatchSnapshot();
 
           expect(await getAllRotations()).toEqual([
             { ...rotations[0], members: ["user-b", "user-c", "user-a"] },
@@ -325,7 +284,9 @@ describe("functions", () => {
             value: "rotate:rotation-not-exist",
           });
           expect(res.body).toEqual({}); // ack
-          expect(getClientCalls("chat.postEphemeral")).toMatchSnapshot();
+          expect(
+            getSlackWebClientCalls("chat.postEphemeral")
+          ).toMatchSnapshot();
 
           expect(await getAllRotations()).toEqual(rotations);
         });
@@ -342,7 +303,7 @@ describe("functions", () => {
             value: "unrotate:rotation-1",
           });
           expect(res.body).toEqual({}); // ack
-          expect(getClientCalls("chat.update")).toMatchSnapshot();
+          expect(getSlackWebClientCalls("chat.update")).toMatchSnapshot();
 
           expect(await getAllRotations()).toEqual([
             { ...rotations[0], members: ["user-c", "user-a", "user-b"] },
@@ -362,7 +323,9 @@ describe("functions", () => {
             value: "unrotate:rotation-not-exist",
           });
           expect(res.body).toEqual({}); // ack
-          expect(getClientCalls("chat.postEphemeral")).toMatchSnapshot();
+          expect(
+            getSlackWebClientCalls("chat.postEphemeral")
+          ).toMatchSnapshot();
 
           expect(
             (await rotationsRef.get()).docs.map((doc) => doc.data())
@@ -381,7 +344,7 @@ describe("functions", () => {
             value: "delete:rotation-1",
           });
           expect(res.body).toEqual({}); // ack
-          expect(getClientCalls("chat.postMessage")).toMatchSnapshot();
+          expect(getSlackWebClientCalls("chat.postMessage")).toMatchSnapshot();
 
           expect(await getAllRotations()).toEqual(
             rotations.filter((r) => r.id !== "rotation-1")
@@ -398,7 +361,9 @@ describe("functions", () => {
             value: "delete:rotation-not-exist",
           });
           expect(res.body).toEqual({}); // ack
-          expect(getClientCalls("chat.postEphemeral")).toMatchSnapshot();
+          expect(
+            getSlackWebClientCalls("chat.postEphemeral")
+          ).toMatchSnapshot();
 
           expect(await getAllRotations()).toEqual(rotations);
         });
@@ -413,7 +378,7 @@ describe("functions", () => {
       await wrappedCron({
         timestamp: "2020-08-08T22:33:44.000Z", // Sun, 09 Aug 2020 07:33:44 JST
       });
-      expect(getClientCalls("chat.postMessage")).toMatchSnapshot();
+      expect(getSlackWebClientCalls("chat.postMessage")).toMatchSnapshot();
       expect(await getAllRotations()).toEqual([
         { ...rotations[0], members: ["user-b", "user-c", "user-a"] },
         { ...rotations[1], members: ["user-q", "user-p"] },
@@ -426,7 +391,7 @@ describe("functions", () => {
       await wrappedCron({
         timestamp: "2020-08-09T22:33:44.000Z", // Mon, 10 Aug 2020 07:33:44 JST
       });
-      expect(getClientCalls("chat.postMessage")).toMatchSnapshot();
+      expect(getSlackWebClientCalls("chat.postMessage")).toMatchSnapshot();
       expect(await getAllRotations()).toEqual(rotations);
     });
   });
